feat: Remove arquivo run.py e integra lógica de inicialização no app.py
